+++ mobile/android/android-components/components/browser/menu/src/main/java/mozilla/components/browser/menu/item/BrowserMenuItemToolbar.kt
@@ -30,27 +30,42 @@ class BrowserMenuItemToolbar(
 
         for (item in items) {
             val button = AppCompatImageButton(view.context)
-            button.setImageResource(item.imageResource)
+            if (item is TwoStateButton && !item.isInPrimaryState.invoke()) {
+                button.setImageResource(item.secondaryImageResource)
+                button.contentDescription = item.secondaryContentDescription
+                if (item.secondaryImageTintResource != 0) {
+                    button.imageTintList = ContextCompat.getColorStateList(
+                        view.context,
+                        item.secondaryImageTintResource
+                    )
+                }
+                button.isEnabled = !item.disableInSecondaryState
+            } else {
+                button.setImageResource(item.imageResource)
+                button.contentDescription = item.contentDescription
+                if (item.iconTintColorResource != 0) {
+                    button.imageTintList =
+                        ContextCompat.getColorStateList(view.context, item.iconTintColorResource)
+                }
+            }
 
             val outValue = TypedValue()
             view.context.theme.resolveAttribute(
-                    android.R.attr.selectableItemBackgroundBorderless,
-                    outValue,
-                    true)
+                android.R.attr.selectableItemBackgroundBorderless,
+                outValue,
+                true
+            )
 
             button.setBackgroundResource(outValue.resourceId)
-            button.contentDescription = item.contentDescription
             button.setOnClickListener {
                 item.listener.invoke()
                 menu.dismiss()
             }
 
-            if (item.iconTintColorResource != 0) {
-                button.imageTintList = ContextCompat.getColorStateList(view.context, item.iconTintColorResource)
-            }
-
-            layout.addView(button,
-                LinearLayout.LayoutParams(0, view.resources.pxToDp(ICON_HEIGHT_DP), 1f))
+            layout.addView(
+                button,
+                LinearLayout.LayoutParams(0, view.resources.pxToDp(ICON_HEIGHT_DP), 1f)
+            )
         }
     }
 
@@ -62,11 +77,42 @@ class BrowserMenuItemToolbar(
      * @param iconTintColorResource Optional ID of color resource to tint the icon.
      * @param listener Callback to be invoked when the button is pressed.
      */
-    class Button(
+    open class Button(
         val imageResource: Int,
         val contentDescription: String,
         val iconTintColorResource: Int = 0,
         val listener: () -> Unit
+    )
+
+    /**
+     * A button that either shows an primary state or an secondary state based on the provided
+     * <code>isInPrimaryState</code> lambda.
+     *
+     * @param primaryImageResource ID of a drawable resource to be shown as primary icon.
+     * @param primaryContentDescription The button's primary content description, used for accessibility support.
+     * @param primaryImageTintResource Optional ID of color resource to tint the primary icon.
+     * @param secondaryImageResource Optional ID of a different drawable resource to be shown as secondary icon.
+     * @param secondaryContentDescription Optional secondary content description for button, for accessibility support.
+     * @param secondaryImageTintResource Optional ID of secondary color resource to tint the icon.
+     * @param isInPrimaryState Lambda to return true/false to indicate if this button should be primary or secondary.
+     * @param disableInSecondaryState Optional boolean to disable the button when in secondary state.
+     * @param listener Callback to be invoked when the button is pressed.
+     */
+    open class TwoStateButton(
+        val primaryImageResource: Int,
+        val primaryContentDescription: String,
+        val primaryImageTintResource: Int = 0,
+        val secondaryImageResource: Int = primaryImageResource,
+        val secondaryContentDescription: String = primaryContentDescription,
+        val secondaryImageTintResource: Int = primaryImageTintResource,
+        val isInPrimaryState: () -> Boolean = { true },
+        val disableInSecondaryState: Boolean = false,
+        listener: () -> Unit
+    ) : Button(
+        primaryImageResource,
+        primaryContentDescription,
+        primaryImageTintResource,
+        listener = listener
     )
 
     companion object {